Initial commit
This is the initial commit of my social network bubble simulation
project. The idea is to model how a social network behaves if there
are dissenting opinions in the network, and the network can be
modified due to dissenting opinions.

This is a rust project with some very basic dependencies. Use cargo
to build. I'll be working on some script automation to run many
simulations to analyze, and in the future I will be working on
different aspects of refactoring the simulation to be a bit more
performant.

Possible sources of performance enhancements include the use of
concurancy for some basic evaluation code, algorithim optimizations,
and the simpliciation of some more complex parts of the model.

I'd also like to work on documenting the simulation application a
bit better so as to make it more maintainable in the future.