Added network and metadata output
This commit adds the ability to write the state of the network to a
file, as well as metadata regarding the simulation that was performed.
This will become very useful once I start performing bulk simulations.
Once I have bulk simulations up and running, I will need to be able to
identify specific simulation parameters reliably.